[ESWE-1217] Remove Example APIs
