+++ helm_deploy/values-prod.yaml
@@ -7,9 +7,6 @@ generic-service:
 
   env:
     HMPPS_AUTH_URL: "https://sign-in.hmpps.service.justice.gov.uk/auth"
-    # Template kotlin calls out to itself to provide an example of a service call
-    # TODO: This should be replaced by a call to a different service, or removed
-    EXAMPLE_API_URL: "https://jobs-board-integration-api.hmpps.service.justice.gov.uk"
     API_BASE_URL_JOBSBOARD: "https://jobs-board-api.hmpps.service.justice.gov.uk"
     API_BASE_URL_MNJOBBOARD: "https://liveservices.sequation.net/sequation-job-api"   # URL to be confirmed
