+++ src/integrationTest/kotlin/uk/gov/justice/digital/hmpps/jobsboardintegrationapi/integration/health/HealthCheckTest.kt
@@ -30,7 +30,7 @@ class HealthCheckTest : IntegrationTestBase() {
       .expectBody()
       .jsonPath("status").isEqualTo("DOWN")
       .jsonPath("components.hmppsAuth.status").isEqualTo("DOWN")
-      .jsonPath("components.exampleApi.status").isEqualTo("DOWN")
+      .jsonPath("components.jobsBoardApi.status").isEqualTo("DOWN")
   }
 
   @Test
