+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboardintegrationapi/config/WebClientConfiguration.kt
@@ -15,31 +15,24 @@ import java.time.Duration
 
 @Configuration
 class WebClientConfiguration(
-  @Value("\${example-api.url}") val exampleApiBaseUri: String,
   @Value("\${hmpps-auth.url}") val hmppsAuthBaseUri: String,
+  @Value("\${api.base.url.jobsboard}") val jobsboardApiBaseUri: String,
   @Value("\${api.health-timeout:2s}") val healthTimeout: Duration,
   @Value("\${api.timeout:20s}") val timeout: Duration,
   @Value("\${api.client.logging.enabled:false}") val clientLogging: Boolean,
 ) {
   // HMPPS Auth health ping is required if your service calls HMPPS Auth to get a token to call other services
-  // TODO: Remove the health ping if no call outs to other services are made
   @Bean
   fun hmppsAuthHealthWebClient(builder: WebClient.Builder): WebClient = builder.healthWebClient(hmppsAuthBaseUri, healthTimeout)
 
-  // TODO: This is an example health bean for checking other services and should be removed / replaced
   @Bean
-  fun exampleApiHealthWebClient(builder: WebClient.Builder): WebClient = builder.healthWebClient(exampleApiBaseUri, healthTimeout)
-
-  // TODO: This is an example bean for calling other services and should be removed / replaced
-  @Bean
-  fun exampleApiWebClient(authorizedClientManager: OAuth2AuthorizedClientManager, builder: WebClient.Builder): WebClient = builder.authorisedWebClient(authorizedClientManager, registrationId = "example-api", url = exampleApiBaseUri, timeout)
+  fun jobsBoardHealthWebClient(builder: WebClient.Builder): WebClient = builder.healthWebClient(jobsboardApiBaseUri, healthTimeout)
 
   @ConditionalOnIntegrationEnabled
   @Bean
   fun jobsBoardWebClient(
     authorizedClientManager: OAuth2AuthorizedClientManager,
     builder: WebClient.Builder,
-    @Value("\${api.base.url.jobsboard}") jobsboardApiBaseUri: String,
   ): WebClient = builder.apply {
     if (clientLogging) it.clientConnector(clientConnectorWithLogging())
   }.authorisedWebClient(
