+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboardintegrationapi/health/HealthPingCheck.kt
@@ -6,10 +6,8 @@ import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.hmpps.kotlin.health.HealthPingCheck
 
 // HMPPS Auth health ping is required if your service calls HMPPS Auth to get a token to call other services
-// TODO: Remove the health ping if no call outs to other services are made
 @Component("hmppsAuth")
 class HmppsAuthHealthPing(@Qualifier("hmppsAuthHealthWebClient") webClient: WebClient) : HealthPingCheck(webClient)
 
-// TODO: Example ping health check calling out to other services
-@Component("exampleApi")
-class ExampleApiHealthPing(@Qualifier("exampleApiHealthWebClient") webClient: WebClient) : HealthPingCheck(webClient)
+@Component("jobsBoardApi")
+class JobsBoardApiHealthPing(@Qualifier("jobsBoardHealthWebClient") webClient: WebClient) : HealthPingCheck(webClient)
