+++ src/main/resources/application-dev.yml
@@ -1,14 +1,6 @@
 hmpps-auth:
   url: "http://localhost:8090/auth"
 
-# example client configuration for calling out to other services
-# TODO: Remove / replace this configuration
-example-api:
-  url: "http://localhost:8080"
-  client:
-    id: "example-api-client"
-    secret: "example-api-client-secret"
-
 api:
   base.url:
     jobsboard: "https://jobs-board-api-dev.hmpps.service.justice.gov.uk"
